Update usage to 64-bit
Bug: 284333373
Test: Mapper5 VTS GetUsage64

+++ cros_gralloc/cros_gralloc_buffer.cc
@@ -115,9 +115,9 @@ int32_t cros_gralloc_buffer::get_android_format() const
 	return hnd_->droid_format;
 }
 
-uint64_t cros_gralloc_buffer::get_android_usage() const
+int64_t cros_gralloc_buffer::get_android_usage() const
 {
-	return static_cast<uint64_t>(hnd_->usage);
+	return hnd_->usage;
 }
 
 int32_t cros_gralloc_buffer::increase_refcount()
